fix(jvim-timer): update JvimTimer.java
- change output from "12.3 секунд" to "12 сек"

version 0.1.0

+++ jvim-timer/java/JvimTimer.java
@@ -20,10 +20,10 @@ public class JvimTimer {
       reader.close();
             
       long duration = System.currentTimeMillis() - startTime;
-      double seconds = duration / 1000.0;
+      long seconds = duration / 1000;
             
       System.out.println("Время работы Vim: " 
-                          + String.format("%.1f", seconds) + " сек");
+                          + String.format("%d", seconds) + " сек");
             
       new File("/tmp/vim_start_time.txt").delete();
